refactor(task): moved task registration logic for GenerateNdkSoMapping into the Task instead of the Plugin
